pyproject.toml
[tool.poetry]
name = "lime"
version = "0.0.1dev0"
description = "A personal job board."
authors = ["Artem Malyshev <proofit404@gmail.com>"]
packages = [
  { include = "app", from = "src" },
  { include = "examples", from = "src" },
]

[tool.poetry.dependencies]
python = "3.8.2"
attrs = "19.3.0"
dependencies = "1.0.1"
dramatiq = {version = "1.8.1", extras = ["rabbitmq"]}
flask = "1.1.2"
generics = "1.1.0"
lxml = "4.5.2"
marshmallow = "3.5.1"
requests = "2.23.0"
stories = "0.14.0"
wtforms = "2.2.1"
xmlschema = "1.2.3"

[tool.poetry.dev-dependencies]
bandit = "1.6.2"
bellybutton = "0.3.0"
betamax = "0.8.1"
black = "19.10b0"
coverage = "5.0.4"
docformatter = "1.3.1"
flake8 = "3.8.1"
flake8-2020 = "1.6.0"
flake8-blind-except = "0.1.1"
flake8-bugbear = "20.1.4"
flake8-builtins = "1.5.3"
flake8-comprehensions = "3.2.2"
flake8-docstrings = "1.5.0"
flake8-eradicate = "0.3.0"
flake8-executable = "2.0.3"
flake8-fixme = "1.1.1"
flake8-mutable = "1.2.0"
flake8-print = "3.1.4"
flake8-pytest = "1.3"
flake8-pytest-style = "1.1.1"
flake8-tidy-imports = "4.1.0"
pre-commit = "2.6.0"
pytest = "5.4.1"
pytest-deadfixtures = "2.2.0"
pytest-randomly = "3.2.1"
pytest-testmon = "1.0.2"
pytest-timeout = "1.3.4"
pytest-watch = "4.2.0"
pyupgrade = "2.7.2"
reorder-python-imports = "2.3.5"
xenon = "0.7.0"
yamllint = "1.24.2"
